feat: implement Pictogram and Tile components with tests; refactor imports and file handling
