+++ src/features/board/BoardView.tsx
@@ -1,7 +1,6 @@
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 import { Grid, type GridItemProps } from "./grid/Grid";
-import { Tile } from "./grid/Tile";
 import { MessageBar } from "./message/MessageBar";
 import { useMessage } from "./message/useMessage";
 import { useMessagePlayback } from "./message/useMessagePlayback";
@@ -9,6 +8,7 @@ import { NavButtons } from "./navigation/NavButtons";
 import { useBoardNavigation } from "./navigation/useBoardNavigation";
 import { SuggestionBar } from "./suggestions/SuggestionBar";
 import { useSuggestions } from "./suggestions/useSuggestions";
+import { Tile } from "./Tile";
 import type { Board, BoardButton } from "./types";
 import { useButtonActivation } from "./useButtonActivation";
