+++ src/features/board/Pictogram.test.tsx
