+++ src/features/board/Pictogram.tsx
