+++ src/features/board/Tile.tsx
