+++ src/shared/utils/file-picker.ts
@@ -1,9 +1,20 @@
-export function openFiles(): Promise<File[]> {
+export interface OpenFilesOptions {
+  accept?: string;
+  multiple?: boolean;
+}
+
+export function openFiles({
+  accept,
+  multiple = false,
+}: OpenFilesOptions = {}): Promise<File[]> {
   return new Promise((resolve) => {
     const input = document.createElement("input");
     input.type = "file";
-    input.multiple = true;
-    input.accept = ".obz,.obf,application/zip,application/json";
+    input.multiple = multiple;
+
+    if (accept) {
+      input.accept = accept;
+    }
 
     input.addEventListener("change", () => {
       resolve(input.files ? Array.from(input.files) : []);
